refactor(api): move repo handlers to new package
