+++ api/repo/delete.go
@@ -1,0 +1,98 @@
+// Copyright (c) 2022 Target Brands, Inc. All rights reserved.
+//
+// Use of this source code is governed by the LICENSE file in this repository.
+
+package repo
+
+import (
+	"fmt"
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+	"github.com/go-vela/server/database"
+	"github.com/go-vela/server/router/middleware/org"
+	"github.com/go-vela/server/router/middleware/repo"
+	"github.com/go-vela/server/router/middleware/user"
+	"github.com/go-vela/server/scm"
+	"github.com/go-vela/server/util"
+	"github.com/sirupsen/logrus"
+)
+
+// swagger:operation DELETE /api/v1/repos/{org}/{repo} repos DeleteRepo
+//
+// Delete a repo in the configured backend
+//
+// ---
+// produces:
+// - application/json
+// parameters:
+// - in: path
+//   name: org
+//   description: Name of the org
+//   required: true
+//   type: string
+// - in: path
+//   name: repo
+//   description: Name of the repo
+//   required: true
+//   type: string
+// security:
+//   - ApiKeyAuth: []
+// responses:
+//   '200':
+//     description: Successfully deleted the repo
+//     schema:
+//       type: string
+//   '500':
+//     description: Unable to  deleted the repo
+//     schema:
+//       "$ref": "#/definitions/Error"
+//   '510':
+//     description: Unable to  deleted the repo
+//     schema:
+//       "$ref": "#/definitions/Error"
+
+// DeleteRepo represents the API handler to remove
+// a repo from the configured backend.
+func DeleteRepo(c *gin.Context) {
+	// capture middleware values
+	o := org.Retrieve(c)
+	r := repo.Retrieve(c)
+	u := user.Retrieve(c)
+
+	// update engine logger with API metadata
+	//
+	// https://pkg.go.dev/github.com/sirupsen/logrus?tab=doc#Entry.WithFields
+	logrus.WithFields(logrus.Fields{
+		"org":  o,
+		"repo": r.GetName(),
+		"user": u.GetName(),
+	}).Infof("deleting repo %s", r.GetFullName())
+
+	// send API call to remove the webhook
+	err := scm.FromContext(c).Disable(u, r.GetOrg(), r.GetName())
+	if err != nil {
+		retErr := fmt.Errorf("unable to delete webhook for %s: %w", r.GetFullName(), err)
+
+		// set the status code based off the
+		code := http.StatusInternalServerError
+		if err.Error() == "Repo not found" {
+			code = http.StatusNotFound
+		}
+
+		util.HandleError(c, code, retErr)
+
+		return
+	}
+
+	err = database.FromContext(c).DeleteRepo(r)
+	if err != nil {
+		retErr := fmt.Errorf("unable to delete repo %s: %w", r.GetFullName(), err)
+
+		util.HandleError(c, http.StatusInternalServerError, retErr)
+
+		return
+	}
+
+	c.JSON(http.StatusOK, fmt.Sprintf("repo %s deleted", r.GetFullName()))
+}
